Allow Nones in Tensor[] for IValues
When we parse IValues from Python to C++, we treat Nones as Tensors
(these get parsed into undefined tensors). Before this PR, this logic
only happened for Tensor (and not Tensor[]). For example, we could
create a custom operator foo(Tensor x) and it would accept None as
input.

This PR makes the behavior consistent by extending it to Tensor[].

I need this behavior later on when parsing lists of grad_output which
look like they are List[Tensor] in C++ but become List[Optional[Tensor]]
in Python (because undefined tensor in C++ maps to Python None).

Test Plan:
- new test

ghstack-source-id: 2b4a4a4d3b270962b2a7429d45a82c83ff6efd66
Pull Request resolved: https://github.com/pytorch/pytorch/pull/143168