Use the device interface for detecting Triton availability
This allows for each device type to check installed devices for Triton
compatibility and ensure their Triton backend is present.

This PR replaces the has_triton() global method which was previously
used for this task, and moves the initial check for each Inductor
backend on to their associated BaseScheduler subclass. This means that
other backends, such as Halide, can also implement their own
availability checks.