Make MultiProcContinuousTest timeout configurable
Allows test classes using MPCT to set their own timeout as a class
property, which is good enough since the processgroup is shared across
test instances and the timeout is set at processgroup init.

Also sets a default timeout of 2 minutes, which is probably (?) long
enough for reasonable tests, but can be changed if it causes flakyness.
It's preferable to have as short default timeout as possible, since when
debugging tests getting a timeout quickly helps.

ghstack-source-id: bdf0c0777eec530e2023a5b4bb28ba9dd2695351
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145099