[dynamo] Fix read/write conflicts in a cuda test
Prior to this patch, the `test_cuda_event_created_outside_of_graph`
is flaky in CI, and that's because we have read and write to the same
`foo` tensor buffer from 2 different streams. This patch eliminates that
by adding a synchronization to wait till read finishes before starting
the write.

Fixes #133837, #133828.

ghstack-source-id: 3f2a2d100161cc6d5fe3cde3358fc697723d098b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145658