Avoid data-dependent errors by runtime assert substitution.
This PR adds a simplification method using runtime asserts, whenever we
are about to raise a data-dependent error. We use the recorded runtime
asserts as a source of knowledge for substituting the free symbols in
the given expression.

This is useful for avoiding data-dependent errors, specifically avoiding
guarding on expressions with unbacked integers, that are deducible from
the past runtime asserts.

ghstack-source-id: c2ccc3b6294980ef57a6a17c6f49d9c2bbd046bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145681