[bug] handle case when remote peer closes connection
Summary:
In the case where remote peer closes the connection, nread returns 0. In
this case, we still want to free up the allocated buffer.
Also, reorder the if so that the likely success cases (nread > 0) is at
the top of the function with an early return.

Test Plan:
unit tests

ghstack-source-id: ceaa081e324acfc0f49e7eb2daea3e8b0ac6332a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145757