[bug] handle case when remote peer closes connection
Summary:
In the case where remote peer closes the connection, nread returns 0. In
this case, we still want to free up the allocated buffer.
Also, reorder the if so that the likely success cases (nread > 0) is at
the top of the function with an early return.

Test Plan:
unit tests

ghstack-source-id: ca00aabbc4b1298e396372df8f0a73d90f99d4fc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145757