bump counters for unbacked binding names
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145882

Instead of bumping symint counters when we process unbacked bindings during deserialization, it's better to bump them at the beginning based on what the symbols in the original shape env before serialization were. This allows symbols in unbacked bindings to have "gaps" that bumping alone would not be able to match.

Why is bumping counters important at all? It is because when the shape env coming out of deserialization is used later for propagating symints, say in run_decompositions, we don't want new names to clash with existing names (bad things happen).
ghstack-source-id: 263554936
@exported-using-ghexport

Differential Revision: [D68798191](https://our.internmc.facebook.com/intern/diff/D68798191/)